infra/docker/requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
boto3>=1.35.0
pyyaml>=6.0
requests>=2.31.0
